Adds --no-nucs option to privent saving extra columns for nucs
Avoids saving extra columns for nucleosomes (chrom1 start1 end1 chrom2 start2 end2) in the generated nuc-nuc interaction matrices. Instead, it counts on nuc_id1 and nuc_id2 (which are generated automatically) to identify nucleosomes. This option reduces the size of the produced matrix files, which can improve efficiency in memory and storage usage and speed.