Multiprocessing platform dependency issues: Linux
This commit provides a workaround for Linux for the following Python bug.

There is a platform dependent bug in multiprocessing module:

https://bugs.python.org/issue38119

As mentioned in the link, this problem affects all "Versions: Python 3.10, Python 3.9, Python 3.8".
The bug "breaks the expected cross-platform behavior that a shared memory segment persists at least as long as any running process has a handle on that segment."