Multiprocessing platform dependency issues: Windows
This commit provides a workaround for Windows becuase of the following Python bug. This means there is no mulitprocessing for Windows for the interactions file splitting step, which makes this step much slower in Windows than in macOS or Linux.

The report for the platform dependent bug in multiprocessing module:

https://bugs.python.org/issue38119

As mentioned in the link, this problem affects all "Versions: Python 3.10, Python 3.9, Python 3.8".
The bug "breaks the expected cross-platform behavior that a shared memory segment persists at least as long as any running process has a handle on that segment."